backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
networkx
python-dotenv
langchain
langgraph
langchain-core
langgraph-checkpoint-redis
langchain-openai